% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useless_functions.R
\name{bad.lm}
\alias{bad.lm}
\title{Bad linear model}
\usage{
bad.lm(y, X0)
}
\arguments{
\item{y}{response variable}

\item{X0}{explanatory variable}
}
\value{
fitted parameters
}
\description{
Finds LS solution of parameters for linear model. Much less efficient than lm command, would not recommend using.
}
