% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useless_functions.R
\name{nothing}
\alias{nothing}
\title{Nothing}
\usage{
nothing(x)
}
\arguments{
\item{x}{Anything}
}
\value{
zero
}
\description{
A function that always returns 0
}
\examples{
nothing(1000)
nothing("nothing")
}
