% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useless_functions.R
\name{same}
\alias{same}
\title{Return input}
\usage{
same(x)
}
\arguments{
\item{x}{Anything}
}
\value{
What you input
}
\description{
Function to return whatever you put in
}
\examples{
same(1)
same("same")
}
