% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRPackage.R
\name{uselessFunction}
\alias{uselessFunction}
\title{uselessFunction: Informs the user of its own self-awareness.}
\usage{
uselessFunction(query = "anything at all")
}
\arguments{
\item{query}{Any quoted string}
}
\description{
Takes any input, returns "I am thinking", then "I am self-aware. Honest". Totes
passes the Turing test.
}
